Divide Post components into post and comment folders
