+++ packages/client/src/modules/post/components/post/PostEditView.web.tsx
@@ -2,12 +2,12 @@ import React from 'react';
 import Helmet from 'react-helmet';
 import { Link } from 'react-router-dom';
 
-import { PageLayout } from '../../common/components/web';
+import { PageLayout } from '../../../common/components/web';
 import PostForm from './PostForm.web';
-import PostComments from '../containers/comment/PostComments.web';
-import settings from '../../../../../../settings';
+import PostComments from '../../containers/comment/PostComments.web';
+import settings from '../../../../../../../settings';
 
-import { PostProps, AddPostFn, EditPostFn, Post } from '../types';
+import { PostProps, AddPostFn, EditPostFn, Post } from '../../types';
 
 const onSubmit = (post: Post, addPost: AddPostFn, editPost: EditPostFn) => (values: Post) => {
   if (post) {
